% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_ml_data.R
\name{gen_multilevel_data}
\alias{gen_multilevel_data}
\title{gen_multilevel_data}
\usage{
gen_multilevel_data(
  n = 5,
  n_j = 5,
  B = c(1, 1),
  tau = c(1, 1),
  sigma = 1,
  r = 0.3,
  type = "categorical"
)
}
\arguments{
\item{n}{A numeric specifying the number of units.}

\item{n_j}{A numeric specifying the number of observations per unit.}

\item{B}{fixed effects (intercept + slope)}

\item{tau}{Random effect SDs}

\item{sigma}{residual SD}

\item{r}{correlation for random effects}

\item{type}{categorical or continuous}
}
\description{
gen_multilevel_data
}
